Clarify that WorldLocation.all doesn't return delegations/missions
I hadn't appreciated that `WorldLocation.all` only returns results
which have a format of "World location" [1].

Furthermore I hadn't appreciated that for location slugs containing either
"delegation" or "mission", `worldwide_api_has_locations` sets the format of
the stubbed locations to "International delegation", not "World location" [2].

I've tweaked this `WorldLocationTest` slightly in an attempt to make the above
a bit clearer. The commit does not change the behaviour of either of the above
methods, but expands the name of the test and adds a mission to the list of
location slugs.

Ideally all the tests in `WorldLocationTest` would stub `Services.worldwide_api`
and not use the `GdsApi::TestHelpers::Worldwide` helper methods at all. If we
were doing this would make it easier to explain the behaviour of
`WorldLocation.all` in a test, but that's a job for another day [3].

I came across this when I was trying to understand why the following locations
were not appearing in the `country_select` questions in
`CountryAndDateQuestionsTest`:

* `uk-delegation-to-council-of-europe`
* `uk-delegation-to-organization-for-security-and-co-operation-in-europe`

[1]: https://github.com/alphagov/smart-answers/blob/c265170b8c94984daa83924bc50b3ec796b2b70d/app/models/world_location.rb#L17
[2]: https://github.com/alphagov/gds-api-adapters/blob/29118ff686ecdb05ea338f6dd788bfa7015d69bd/lib/gds_api/test_helpers/worldwide.rb#L100
[3]: https://trello.com/c/M4Y09qFy