Remove delegations from CountryAndDateQuestionsTest
The fact that `GdsApi::Worldwide::TestHelpers#worldwide_api_has_locations`
sets the format of locations with a slug containing the word "delegate" to
"International delegation" and the fact that `WorldwideLocation.all` filters
these out is sufficiently well tested in `WorldwideLocationTest`.

Testing this kind of detail in an integration test doesn't seem very sensible.

I found the test very hard to understand i.e. why were the delegations not
showing up in the select control on the page.

Thus I'm happy to remove the delegation locations from the test setup and as I
expect none of the assertions need to change.