Generate more realistic names for WorldLocation stubs
I want to use these stubbing methods in the regression tests, but previously
the generated names did not match those in the regression test artefacts. This
change means that the generated name will be exactly the same as if we'd been
using the `GdsApi::TestHelpers::Worldwide` helper methods.

Note that the generated names are still not that great for slugs including the
word "and", because the "and" gets capitalized along with the other words e.g.
"st-pierre-and-miquelon" -> "St Pierre And Miquelon".

However, since this is how the regression test artefacts have been generated,
it seems the most sensible option for now.