Select type of diplomatic mission.
This adds test and implementation to help decide which diplomatic mission a country should have.

From my (prior) knowledge, High Commissions are setup in Commonwealth countries and
embassies are setup in none common wealth countries.

For this card, I have focused on settling only the Australian (same-sex) and the other countries scenario. This will allow for this card to move on
whilst more clarity on this matter is obtained and implemented.