Move formatting of holiday_entitlement_days into templates
The value of `holiday_entitlement_days` is calculated in two different
`calculate` blocks. The first one returns a number (mostly integers, but one
float value of 7.5). The second one returns a number formatted as a string,
with a precision of 1 i.e. always displaying 1 decimal place. This means:

* an integer from the 1st block will be displayed with 0 decimal places
* a float from the 1st block will be displayed with all its decimal places
* a float from the 2nd block will always be displayed with 1 decimal place

Note that the `sprintf` formatting in the 2nd block also does some rounding, but
not quite as I would expect:

* `sprintf(''%.1f', 0.25)` # => 0.2
* (0.25).round(1)          # => 0.3

I'm pretty sure this formatting/rounding is not working as it was intended and
it could certainly result in inconsistent values being displayed if for example
the first block started returning floats with more than one decimal place. I've
created a Trello card to capture this potential problem [1].

Also I'm not entirely sure what the point of the calls to `round(10)` are meant
to be achieving in both of the `calculate` blocks. However, these calls seem to
to permeate the code more extensively than just this flow, so I've created
another Trello card to investigate [2].

Putting all the above to one side, this commit creates a new (almost duplicate)
outcome so that we can move the number formatting into the relevant outcome
templates while retaining the odd behaviour described above. I know this isn't
ideal, but I didn't want to get distracted into making a change that would
probably need fact-checking.

I'd have preferred to have used a standard helper method like
`ActionView::Helpers::NumberHelper` [3], e.g.
`number_with_precision(days, precision: 1, strip_insignificant_zeros: true)`,
but there wasn't an easy way to do this without changing existing the behaviour.

Although there is no change to the external behaviour, I've updated the
regression test responses & expected results to capture the new outcome and I've
added the new outcome template to the checksums file. All the regression tests
are passing and so I'm happy to update the checksums (some of which are due to
changes in previous commits).

I've also updated a couple of assertions in the integration test to match the
changes in internal behaviour.

I've also had to add an override for the `MethodMissingHelper` in order to keep
the `IndexableContentTest` passing.

[1]: https://trello.com/c/HXgHxbpp
[2]: https://trello.com/c/0WNxygY4
[3]: http://api.rubyonrails.org/classes/ActionView/Helpers/NumberHelper.html#method-i-number_with_precision