Remove redundant to_i conversion in ChildMaintenanceCalculator
The response to the `how_many_children_paid_for?` question is already being
converted to an integer in the `number_of_children` calculate block so there's
no need to do it here as well.

The conversion is a user interface concern and so doesn't belong in the "model"
in any case.