Inline paying_or_receiving_hint into question template
This is a presentational concern and so belongs in the question template. This
also has the advantage that we don't need to call `html_safe` on the text.

I've deleted the assertion in the integration test that was checking the hint
text. I don't think this was adding much value, especially given that the hint
is captured in the regression test artefact [1].

[1]: `test/artefacts/calculate-your-child-maintenance/pay.html