Inline rate_type_formatted into outcome template
This is a presentational concern and so it belongs in the outcome template.

Inspecting `ChildMaintenanceCalculator#rate_type` and the associated data stored
in `child_maintenance_data.yml`, I can see that the possible return values are:

* `:nil`
* `:flat`
* `:reduced`
* `:basic`
* `:basic_plus`

The only special case in the `rate_type_formatted` block was to remove the
underscore from `:basic_plus` - the only value which includes an underscore.

Considering this and the sentence in which `rate_type_formatted` is used, it
seems pretty obvious that the intention is to convert the symbols into something
suitable for human consumption and thus I was happy to use `String#humanize` to
do the conversion in the template.

I've removed an assertion from the integration test. I don't think this was
serving much purpose and in any case it's covered in the integration tests.