Remove redundant if from holiday_entitlement_days calculate block
The `holiday_entitlement_days` state variable is only used in the two outcome
templates. However, in the case where the response to the
`worked_for_same_employer?` question is "multiple-employer", the state variable
is over-written by a calculate block in the
`how_many_weeks_at_current_employer?` question. So this `if` statement is
redundant and can be removed.