Render fees partial in outcome for Dominican Republic
Note that the logic in the fees partial is slightly different from the logic in
the template. The fees partial only displays the link to consular fees and the
payment information for countries that aren't in the
`COUNTRIES_WITHOUT_CONSULAR_FACILITIES` array (i.e.
`calculator.country_without_consular_facilities?` returns `false`). The
Dominican Republic isn't in this array so the overall behaviour is unchanged in
this commit.