Render fees partial in same sex consular countries outcome
I used `script/marriage-abroad-countries-leading-to-outcome.rb` (with _all_
countries added to the marriage-abroad regression tests) to determine the
following countries that lead to this outcome:

* bulgaria
* croatia
* cyprus
* guatemala
* moldova
* panama
* venezuela

Note that the logic in the fees partial is slightly different from the logic in
the template. The fees partial only displays the link to consular fees and the
payment information for countries that aren't in the
`COUNTRIES_WITHOUT_CONSULAR_FACILITIES` array (i.e.
`calculator.country_without_consular_facilities?` returns `false`). None of the
7 countries listed above are in this array so the overall behaviour is
unchanged in this commit.