Render fees partial in outcomes for Kosovo
Note that the logic in the fees partial is slightly different from the logic in
the templates. The fees partial only displays the link to consular fees and the
payment information for countries that aren't in the
`COUNTRIES_WITHOUT_CONSULAR_FACILITIES` array (i.e.
`calculator.country_without_consular_facilities?` returns `false`). Kosovo
isn't in this array so the overall behaviour is unchanged in this commit.