Render fees partial in outcome for Romania
Note that the name of this outcome template doesn't follow the pattern of
naming that we're using for all other templates. I've created a Trello card to
remind us to rename it: https://trello.com/c/XG0foVvN.

Note that the logic in the fees partial is slightly different from the logic in
the template. The fees partial only displays the link to consular fees and the
payment information for countries that aren't in the
`COUNTRIES_WITHOUT_CONSULAR_FACILITIES` array (i.e.
`calculator.country_without_consular_facilities?` returns `false`). Romania
isn't in this array so the overall behaviour is unchanged in this commit.