Render fees partial in marriage outcome for France and territories
I used `script/marriage-abroad-countries-leading-to-outcome.rb` (with _all_
countries added to the marriage-abroad regression tests) to determine the
following countries that lead to this outcome:

* france
* french-guiana
* french-polynesia
* guadeloupe
* martinique
* mayotte
* new-caledonia
* reunion
* st-pierre-and-miquelon
* wallis-and-futuna

I've added these countries to the conditional in the fees partial so that we
don't display the link to the consular fee or the payment information. Note
that France, New Caledonia and Wallis and Futuna were added to the conditional
in the previous commit.