Render fees partial in outcome for same sex marriage and civil partnership
I used `script/marriage-abroad-countries-leading-to-outcome.rb` (with _all_
countries added to the marriage-abroad regression tests) to determine the
following countries that lead to this outcome:

* albania
* australia
* azerbaijan
* bolivia
* cambodia
* chile
* china
* colombia
* estonia
* germany
* hungary
* japan
* kosovo
* latvia
* lithuania
* mongolia
* montenegro
* nicaragua
* peru
* philippines
* russia
* serbia
* vietnam

Note that the logic in the fees partial is slightly different from the logic in
the template. The fees partial only displays the link to consular fees and the
payment information for countries that aren't in the
`COUNTRIES_WITHOUT_CONSULAR_FACILITIES` array (i.e.
`calculator.country_without_consular_facilities?` returns `false`). None of the
countries listed above appear in this array so the overall behaviour is
unchanged in this commit.

The fees partial already contains the conditional logic for Vietnam and
Cambodia which is why that partial remains unchanged in this commit.