Render fees partial in opposite sex consular cni countries outcomes
'cote-d-ivoire' is already in the COUNTRIES_WITHOUT_CONSULAR_FACILITIES array
so it was unnecessarily in the conditional statement in the template. Hence my not adding
it to the list of excluded countries in the fees partial.

I've added a condition to the fees partial to ensure that we don't display the
link to consular fees for ceremonies in Brazil when resident of the UK or
ceremony country. This mirrors the behaviour in the two outcome templates
updated in this commit.

I used `script/marriage-abroad-countries-leading-to-outcome.rb` (with _all_
countries added to the marriage-abroad regression tests) to determine the
following countries that lead to these two outcomes:

* afghanistan
* albania
* algeria
* american-samoa
* andorra
* angola
* armenia
* aruba
* austria
* azerbaijan
* bahrain
* benin
* bhutan
* bolivia
* bonaire-st-eustatius-saba
* bosnia-and-herzegovina
* brazil
* bulgaria
* burkina-faso
* burundi
* cape-verde
* central-african-republic
* chad
* chile
* comoros
* congo
* croatia
* cuba
* curacao
* democratic-republic-of-congo
* djibouti
* el-salvador
* equatorial-guinea
* eritrea
* estonia
* ethiopia
* finland
* gabon
* greece
* guatemala
* guinea
* guinea-bissau
* haiti
* honduras
* hungary
* iceland
* iraq
* jordan
* kazakhstan
* kyrgyzstan
* latvia
* libya
* liechtenstein
* lithuania
* luxembourg
* macedonia
* mali
* marshall-islands
* mauritania
* mexico
* micronesia
* moldova
* montenegro
* nepal
* nicaragua
* niger
* norway
* palau
* panama
* paraguay
* romania
* russia
* san-marino
* sao-tome-and-principe
* serbia
* south-sudan
* st-maarten
* sudan
* suriname
* sweden
* tajikistan
* timor-leste
* togo
* tunisia
* turkmenistan
* uzbekistan
* venezuela
* western-sahara

Note that the logic in the fees partial is slightly different from the logic in
the templates. The fees partial only displays _link_to_consular_fees and
_payment_information for countries that aren't in the
`COUNTRIES_WITHOUT_CONSULAR_FACILITIES` array (i.e.
`calculator.country_without_consular_facilities?` returns `false`). The
templates display the _link_to_consular_fees partial for countries that aren't
in the `COUNTRIES_WITHOUT_CONSULAR_FACILITIES` array but always display the
_payment_information partial. Despite this difference, there is no change in
behaviour because none of the countries in the
`COUNTRIES_WITHOUT_CONSULAR_FACILITIES` array have data in the
marriage_abroad_services.yml file (i.e. if
`#country_without_consular_facilities?` is `true` then the _payment_information
partial won't render anything anyway).