Inline late_payment_penalty precalculate block into template
Note that I've had to change the state variable assertions in the integration
test to assert against the `late_payment_penalty` method on the calculator.

Note also that one of the occurrences of `late_payment_penalty` in the `late`
outcome template was already using the calculator method, but another was not!