Track transit from the UK to Common Travel Areas
The current state of check uk visa smart-answer doesn't support or
provide information to users who are coming to the UK on transit to
the Channel Islands, Isle of Man or Republic of Ireland (known collectively
as the ‘common travel area’ or ‘CTA’).

Upon request from the Content Team (via this logic document
https://docs.google.com/document/d/1HRO23JljpHboQOzFQ6gSc2ks7fRfpDaXjEEpR0iOkx4/edit),

I have added a new question (2a) and outcome (14a) to the transit branch
of this smart answer.

I have added/used "afghanistan" as the country used to test the new transite node.

This decision was sub-conciously made as it was mentioned a lot during meetings with Tom.

I) Defineid travelling to cta answer

   a) This holds the response/answer from the user for question 2a
   b) Add tests for this attribute

II) Defined question 2a and outcome 14a
    a) Add flow tests to properly track the flow of the transit branch
    b) Implementation of the new transit branch
    c) Add/Fix failing tests related to transits within the flow/integration
       tests

These are changes as requested by the Content team