Regenerate regression test artefacts (check-uk-visa)
This commit shows the affected/changes to regression test artefacts
as a result of adding the new question 2a and outcome 14a

I have reviewed the changes to the regression test artefacts and it is as expected.

Tom and I also carried out extensive validation on the browser and found the responses and outcome as expected.

I have also included integration tests to further test navigates to the expected outcomes/nodes.