Store circumstances on calculator attribute in on_response blocks
Note that there are two occurrences of this in Q2 & Q2A:

* `what_are_your_circumstances?`
* `what_are_your_circumstances_without_bills_help?`

But the user is only ever asked *one* of these questions as per the logic in the
`next_node` block in the `what_are_you_looking_for?` (Q1).

Thus it seems OK to use the same calculator attribute method for both.

I've also defaulted the new attribute to an empty array with the idea that it
will always return an array.