Regenerate regression test artefacts (check-uk-visa)
This commit shows the affected/changes to regression test artefacts
as a result of Content changes to Question 3.

I have reviewed the changes to the regression test artefacts and it is as expected.

Tom and I also carried out extensive validation on the browser and found the responses and outcome as expected.

Integration tests already exist that passed, give assurance that the changes
have not affected flow.