Use benefits_claimed attribute in all which_benefits? blocks
Note that I've had to change the logic a bit to cope with the fact that
`calculator.benefits_claimed` is an array i.e. it is the split version of the
response not the raw response which is a comma-separated string.

For example, I believe the following two conditionals are exactly equivalent:

    if %w(working_tax_credit).include?(response)

    if calculator.benefits_claimed == %w(working_tax_credit)

Thus `benefits_claimed` must only contain the `working_tax_credit` option. If
`working_tax_credit` is not the only option selected, then both versions of the
condition will return `false`.

In the case of the `disabled_or_have_children_question` block, I decided not to
make the equivalent changes to the logic in this commit. This was because
there's so _much_ logic, I thought it would be hard to follow. I plan to address
this in a later commit.