Display payment information independently to consular fee link
The data in the marriage_abroad_services.yml file represents the truth as to
whether we should display the payment information section. As such, we can
simplify the logic by moving it out of the conditional used to determine
whether to display the link to consular fees.

We were previously only rendering the payment_information partial for countries
that don't appear in the `COUNTRIES_WITHOUT_CONSULAR_FACILITIES` array. The
payment_information partial renders content based on the data in
marriage_abroad_services.yml. None of the countries in
`COUNTRIES_WITHOUT_CONSULAR_FACILITIES` have an entry in
marriage_abroad_services.yml so the change in this commit doesn't affect the
behaviour of the system.