Store features on calculator attribute in on_response blocks
Note that this state variable was set in three questions which are on three
independent paths through the flow. Given that the variable is used quite a bit
in the outcome templates, it was simpler to update all three questions in the
same commit.

Note also that I've added a default value of an empty array for the `features`
attribute.