Extract age_variant calculator method
For some reason this is sometimes called when the `date_of_birth` calculator
attribute is not set. This was resulting in an error. In order to sort this out
I've added unit tests around the method and introduced a default value for the
`date_of_birth` attribute which gives the expected return value of `nil` for
the `age_variant` calculator method.

Since the `age_variant` state variable wasn't being surfaced for every template
I decided it was simpler to change all the places it was being used to use the
new calculator method.