Extract measure_help? calculator method
Note that I chose to add a trailing question mark to the method name, because
the return value is *used* as a boolean, even though the implementation returns
either a symbol or `nil`. The trailing question mark is not possible on
`calculate` blocks.

I've also introduced some unit tests around the new method to illustrate that it
returns a truth-y or false-y value.