Extract incomesupp_jobseekers_2 methods onto calculator
These two `calculate` blocks both set the same state variable. I have checked
and it is possible to reach the 2nd `calculate` block (in Q5) after the 1st
`calculate` block (in Q4) has been evaluated. In this scenario the value of the
state variable calculated in the 1st block will be over-written by the value
calculated in the 2nd block.

I wanted to extract a single `incomesupp_jobseekers_2` method onto the
calculator, but because of the above, it's not completely obvious how the logic
in the two `calculate` should be combined. By extracting two separate methods
like this I hope to be able to move step-by-step towards a single method.