Extract disabled_or_have_children_question? calculator method
Note that I chose to add a trailing question mark to the method name, because
the method returns a boolean. The trailing question mark is not possible on
`calculate` blocks.

I've also added fairly exhaustive unit tests for the new calculator method,
because I want to refactor it in subsequent commits.

Note that these unit tests are not intended to endorse the current behaviour,
only to characterise it.