Combine incomesupp_jobseekers_2 method into single method
It was a bit hard to work out under what circumstances we should use the logic
in `incomesupp_jobseekers_2_part_1` and when we should override that with the
logic in `incomesupp_jobseekers_2_part_2`.

The best I could come up with is that we should use the former if the
`disabled_or_have_children?` question (Q5) has not yet been answered and the
latter if it has been answered.

All the tests including the regression tests pass, so I'm happy to make this
change. I've also added unit tests around the `incomesupp_jobseekers_2`
calculator method to explain the idea behind this change.