Extract measure_help_and_property_permission_circumstance? calculator method
Note that this logic was duplicated in three `next_node_calculation` blocks
which are now all redundant and can be removed.

Note that I chose to add a trailing question mark to the method name, because
the method returns a boolean. The trailing question mark is not possible on
`next_node_calculation` blocks.