Extract under_green_deal methods onto calculator
These three `precalculate` blocks both set the same state variable for three
different outcomes. It ought to be possible to encapsulate the policy rules for
when something comes under the Green Deal in a single version of this method. I
assume that the reason the logic is different it an "optimisation" in that it's
only possible to arrive at each of the outcomes if certain conditions have been
met by previous responses. Unfortunately this is going to make it hard to
reverse engineer a single method. However, this is a small step in that
direction.