Combine under_green_deal calculator methods into a single method
This is a first stab at combining these methods into a single method.

Unfortunately just combining the three methods using logical ORs doesn't work,
because the logic in each method is missing some implicit logic which is
effectively implemented by the routing logic i.e. you can only get to a certain
outcome if certain conditions are true.

Thus I've had to reproduce the routing logic which is common to the three final
questions, Qa, Qb & Qc, in order to work out which "part" method to use in the
new single uber-method.

Unfortunately this has resulted in some pretty complex logic. However, doing so
means we will be able to remove the last of the `precalculate` blocks from the
flow. Also I'm *sure* it would be possible to simplify the logic considerably
if we could see the policy rules summarised somewhere.