Regenerate regression test artefacts (Marriage abroad)
These changes are as a result of the introduction of the following partials.

* translation_in_germany
* what_you_need_to_do_germany

The changes are as expected.

For the ceremony and third country artefacts:

 * test/artefacts/marriage-abroad/germany/ceremony_country/partner_british/opposite_sex.txt
 * test/artefacts/marriage-abroad/germany/ceremony_country/partner_local/opposite_sex.txt
 * test/artefacts/marriage-abroad/germany/ceremony_country/partner_other/opposite_sex.txt
 * test/artefacts/marriage-abroad/germany/third_country/partner_british/opposite_sex.txt
 * test/artefacts/marriage-abroad/germany/third_country/partner_local/opposite_sex.txt
 * test/artefacts/marriage-abroad/germany/third_country/partner_other/opposite_sex.txt

I have inspected each file and found the content still remains the same as before.

For the UK:

 * test/artefacts/marriage-abroad/germany/uk/partner_british/opposite_sex.txt
 * test/artefacts/marriage-abroad/germany/uk/partner_local/opposite_sex.txt
 * test/artefacts/marriage-abroad/germany/uk/partner_other/opposite_sex.txt

I have inspected each file and found that the only change occurred with the translation urls.

This is as expected as we are already being asked to replace the lawyers links with translation ones.