Add unit tests for incomesupp_jobseekers_1 calculator method
These tests just characterise the existing behaviour - they are not intended to
endorse that behaviour. I wouldn't be surprised if some of this logic is not as
was intended e.g. always returning `nil` when more than option is selected even
though all the options on their own would've retuned `incomesupp_jobseekers_1`.