Convert disabled_or_have_children from String to Array
The comma-separated String `response` from other checkbox questions is split up
into an array before it is stored on the calculator. This commit brings the
`disabled_or_have_children?` question (Q5) into line with the other ones in the
flow.

Note that I'm still not convinced that the behaviour of the
`incomesupp_jobseekers_1` calculator method is as it was intended e.g. always
returning `nil` if more than one option is selected. However, I don't want to
get into this yet.