Return boolean from incomesupp_jobseekers_2_part_1 calculator method
The return value from this calculator method is always used as a boolean and so
I think it's clearer to change the implementation so that it always returns
`true` or `false` i.e. there doesn't seem any benefit in returning the
`incomesupp_jobseekers_2` symbol.

Note that I've had to update the unit tests to match this new behaviour, but
hopefully it's clear that I've replaced all instances of `assert_nil` with
`refute` (for `false` return value), and all instances of
`assert_equal :incomesupp_jobseekers_2` with `assert`.

I've also added a trailing question mark to the method name to make this
behaviour more obvious.