Regenerate regression tests artefacts (Marriage abroad)
This commit shows the affected/changes to regression test artefacts as a result of removing
the hardcoded fees/payment information on the Brazilian marriage abroad pages.

I have reviewed the changes to the regression test artefacts and it is as expected.

The changes affects opposite sex and same sex outcomes where ceremony country is Brazil and user lives in third country or uk

For third country the fees table appears while where the user live in the uk the fees table or payment information don't appear

Already existing integration tests passed, giving assurance that the aforementioned changes have had no (interfering) affect to the flow.