Update marriage-abroad denmark/ceremony_country same-sex test artefacts
As per the changes in the Google Doc[1].

Copy the equivalent opposite-sex outcome, leave the title as "Civil
partnership" and add the synonyms of civil partnership section.

This results in 3 failing regression tests for paths containing
"denmark/ceremony_country/*/same_sex. There are now 6 failing regression tests
as the previous commit also introduced 3.

[1]: https://docs.google.com/document/d/1QMLTHznjIMN7__cMhlp_4GCrTMa5MNgkUFHaOMLu04U/edit