estimate-self-assessment-penalties: Add an option for 2014-2015 tax year
The GOV.UK service for estimating the penalty for late submission of an Self
Assessment return or late payment of an Self Assessment bill is only set up for
returns for the following tax years:

6 April 2011 to 5 April 2012
6 April 2012 to 5 April 2013
6 April 2013 to 5 April 2014

Updated to also assess penalties for returns for the 2014/2015 tax year.
The deadline for April 2014 to April 2015 returns is 31st January 2016.

Added tax year 2014/2015 to the integration tests