Introduce Denmark specific partial
Copied from
outcome_opposite_sex_marriage_in_consular_cni_countries_when_residing_in_uk_or_ceremony_country.

This outcome is used for opposite sex ceremonies when the user is resident of
UK or Denmark. I know that both of these outcomes need to be changed, and that
we'll also need to use this same outcome for same sex ceremonies. I'll rename
it at the point it's also used for same sex ceremonies.

I used the following command to determine which template is being used for
residents of UK and Denmark:

    $ rails r script/marriage-abroad-query-responses-and-expected-results.rb \
    denmark uk

The next step will be to remove the non-Denmark related code from the template.

I've updated the responses-and-expected-results and the regression tests are
passing in this commit.