Make changes for same sex ceremonies in Denmark
This fixes the 6 failing regression tests introduced in the previous 2 commits.

I've renamed the outcome to indicate that it's no longer only used for opposite
sex ceremonies.

I've updated the marriage_abroad_services.yml file so that the same sex
services when resident in Denmark match their opposite sex counterparts. Note
that I had to change the `same_sex/default` service to `third_country` to avoid
these services showing up on denmark/uk outcomes.

I've updated the _contact_method partial to add a special case for Denmark.
This is because we now display the same contact information on opposite and
same sex outcomes when the user is resident of the UK or Denmark.

The regression tests pass in this commit.