Prepare for splitting the outcome for Dominican Republic
The current outcome for ceremonies in the Dominican Republic
(outcome_ceremonies_in_dominican_republic) contains two logical outcomes:

* One for opposite sex ceremonies
* One for same sex ceremonies

The outcome body contains a single `if/else` statement that decides which
logical outcome to display. Splitting these logical outcomes and moving the
"routing logic" to the flow is more consistent with the majority of this Smart
Answer.

This change will make it easier to split the outcome in a subsequent commit.