Regenerate regression tests artefacts (PIP)
This commit shows the affected/changes to regression test artefacts as a result of the Content

changes from the commit before this.

I have reviewed the changes to the regression test artefacts and it is as expected.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no effect.