Prepare for splitting marriage-abroad outcome for Brazil
The current outcome for ceremonies in Brazil
(outcome_marriage_in_brazil_when_residing_in_brazil_or_third_country) contains
two logical outcomes:

* One for residents of Brazil
* One for residents of a third country (i.e. not resident in the UK or Brazill)

The outcome body contains a single `if/else` statement that decides which
logical outcome to display. Splitting these logical outcomes and moving the
"routing logic" to the flow is more consistent with the majority of this Smart
Answer.

This change will make it easier to split the outcome in a subsequent commit.