Extract registration_country calculator method
Note that I've had to instantiate an instance of `RegistrationsDataQuery` to be
used within the calculator. My plan is to hide the use of this class inside the
calculator and so eventually remove the one instantiated at the start of the
flow definition.

Also I've left the `registration_country` calculate block in place, so that I
don't need to make too many changes in this commit, particularly as there is a
later block with the same name that relates to a different country response.