Store married_couple_or_civil_partnership as attribute on calculator
And use married_couple_or_civil_partnership attribute vs response in blocks
within the same question.

Note that I'm not very convinced this is a good name for the attribute, because
the question text says:

> Were you married to the other parent when your child was born?

i.e. it doesn't mention a civil partnership. However, I've captured this
possible oversight in a new Trello card [1].

[1]: https://trello.com/c/hU8tf898