Extract paternity_declaration? calculator method
Note that I've added a trailing question mark to this method name, because it
returns a truth-y or false-y value.

I'm not very convinced this is a good name for this method, but it's simplest
to keep the same name as the original `next_node_calculation` block. A better
name might be `paternity_declaration_necessary?`.