Prepare to split marriage-abroad outcome for Saudi Arabia
The current outcome template (outcome_opposite_sex_marriage_in_saudi_arabia)
contains two logical outcomes:

* One for residents of Saudi Arabia
* One for residents outside of Saudi Arabia

The outcome body contains a single `if/else` statement that decides which
logical outcome to display. Splitting these logical outcomes and moving the
"routing logic" to the flow is more consistent with the majority of this Smart
Answer.

This change will make it easier to split the outcome in a subsequent commit.