Extract currently_in_north_korea? calculator method
I strongly suspect this method contains a bug, because the
`currently_in_north_korea` `next_node_calculation` block does not have a
`response` argument passed in. Thus the use of `response` inside the block is
calling `State#response` which at this point is always `nil`.

I've written this probable bug up as a Trello card [1] and asked for
confirmation. However, in this commit I've retained the existing behaviour and
added a `TODO` comment to highlight the probably bug.

[1]: https://trello.com/c/U58y0ZdJ