Sort countries alphabetically in question an responses
Sort the list of countries in the questions and responses file for Overseas
passport. It makes it easier to add new countries or check if a country has
been included in the regression tests.

The following changes are expected as a result of running the regression
tests again as the first question for some paths has changed.

Changes in outcome page artefacts have been detected:
 D test/artefacts/overseas-passports/france.html
 D test/artefacts/overseas-passports/france/renewing_new.html
 D test/artefacts/overseas-passports/france/renewing_old/adult.html
?? test/artefacts/overseas-passports/afghanistan.html
?? test/artefacts/overseas-passports/afghanistan/renewing_new.html
?? test/artefacts/overseas-passports/afghanistan/renewing_old/adult.html