Use registration_country vs country_of_birth calculator method
The former will return the latter if no `current_country` is set. The only way
to arrive at this outcome is from the `where_are_you_now?` question (Q5). This
route bypasses the `which_country?` question (Q6) which is the only place which
sets `current_country` and so it is safe to use `registration_country` in this
block, thus making it identical to the other two `overseas_passports_embassies`
blocks.