Update countries list for overseas-passports questions and responses
According to the changes from document [1], and in order to validate
the changes are about to be introuced, the following countries are added
to the list of questions and responses of overseas-passports:

- Georgia
- Russia
- Tunisia

In the same go, the artefacts have been regenerated. I have confidence
that they are correct because no change has been added to the current
tool yet.

[1]: https://docs.google.com/document/d/1I-s8kOeguhIaIgJQ5UKhJ5rIAuKtERPPXaj_P2r7q6M/edit