Extract fee-related methods into calculator
I want to remove references to the `RegistrationsDataQuery` instance created as
a local variable at the top of the flow definition and have the calculator
delegate to its own instance in order to hide the implementation from the flow.

This change made the `reg_data_query` precalculate block for the
`north_korea_result` outcome redundant and so I have removed it.