Update refactoring docs to reflect use of on_response blocks
In #2408 we introduced `on_response` blocks with the idea of using them to build
and store a calculator object in a `calculator` state variable, and to manually
store the response to each question on the calculator. This is instead of using
`save_input_as` which stores the response as a state variable.