Regenerate test artefacts Libya (register-a-death)
Selecting Libya from the list of countries now takes you straight to an
outcome, so most of the test artefacts need to be removed.

The test artefacts for North Korea have also been updated. HTML files
in the test artefacts are created the first time a question is reached.

When the questions in the Libya flow were removed, the question "Where
are you now" was no longer being asked.

The first time this question is asked in the tests now is when North
Korea is selected.

This is why the test artefacts for North Korea have also been
regenerated even though the change itself was limited to Libya.