Add support to retrieve fees from a YML file
Extend the `passport_and_embassy_data_query.rb` to retrieve passport fees
from a YML file.

The YML file have this format:

```
---
adult:
  standard:
    fee: 1
    total: 2
  jumbo:
    fee: 3
    total: 4
```

And the data can be retrieved by:
```
@query.find_passport_fees('adult', 'standard', 'total') # => 2
```

This format for the YML file will be evolve once exception to the default
values are introduced. But the first step is to have all the tool using
this YML file for the default values.

This is done in order to decouple the fees values from the view, as they
are currently hardcoded and are not easy to maintain.