Retrieve fee from YML file: ips1, default, adult, standard
Retrieve the default fee for an adult, with an ips_number equal to 1, 
and a standard 32 pages passport.

## About formatting

When formatting the fee retrieved from the YML we are not using the
formatting helper: `SmartAnswer::FormattingHelper#format_money`. 
This is because the helper function is forcing 0 decimals for whole 
numbers (integers): so 83 would be formatted to £83; 
In the current implementation this difference is not being considered,
so 83 is currently formatted to £83.00.

If it is accepted to use 0 decimals for whole numbers, in incoming 
commits I would be good to refactor so the views use the helper function
to format money, as this would be the ideal scenario.