Extract fees into YML when replacing a ips1 passport
It extracts the fees when replacing passport when the ips_application
is 1 and the docs number is 1.

## Some notes:

1. The templates `_costs.govspeack.erb` and `ips_application_result_online.govspeak.erb`
have duplications; I am not tackling this refactoring here as I want to 
focus on extracting the fees into the YML file.

2. The method: 
`calculator.passport_fee('replace','ips_application_1','ips_docs_number_1','adult', 'standard', 'fee')` 
feels too verbose, and would need to be simplified in incoming commits. 
I am not tackling these changes here as I am focusing on extracting
the fees into the XML file.