New partial template boilers and insulation
The section on boilers and insulation appears in four outcome
templates, and sometimes in multiple places in the templates depending
on the rules.

Created a partial for the boilers and insulation section to:
* Avoid duplication
* Make the outcome templates cleaner and easier to read

Replace the boilers and insulation section in all outcomes with the new
partial template.

No test artefacts were modified by this refactor.