New partial template boilers and insulation
The section on boilers and insulation appears in four outcome
templates, and sometimes in multiple places in the templates depending
on the rules.

Created a partial for the boilers and insulation section to:
* Avoid duplication
* Make the outcome templates cleaner and easier to read
* Streamline the logic for boilers and insulation across the outcomes

Also, eventhough the same partials are rendered in the same order in
the outcomes, the rules for rendering the partials can be slightly
different.

As far as I can tell, there should be no reason for these differences:

1. Draught proofing
One outcome displays the draught proofing text if Draught proofing
hasn't been selected in the previous question.

In the other three outcomes, Mains gas has to not have been selected
(deselected) as well.

As there is a clear majority, I have used the rule that states that
either main gas or draught proofing has to not have been selected for
the draught proofing text to be displayed..

2.  Cavity wall and Solid wall insulation

All four outcome templates can display the Cavity wall insulation and
Solid wall insulation, and each one of them does it differently

a. outcome_bills_and_measures_on_benefits_not_eco_eligible.govspeak.erb:
If main gas is not selected, the bullets on cavity wall and solid wall
insulation are displayed, even if the user has indicated that they
already have these two things.

b. outcome_measures_help_deal.govspeak.erb:
The cavity wall and solid wall insulation bullet points are always
displayed, regardless of things selected in the previous question.

c. outcome_bills_and_measures_no_benefits.govspeak.erb:
Cavity wall insulation is displayed if main gas has not been selected
in the previous question. It doesn't check if cavity wall insulation was
selected.

Solid wall insulation is displayed if either main gas or solid wall
insulation was not selected in the previous question.

d. outcome_bills_and_measures_on_benefits_eco_eligible.govspeak.erb:
Cavity wall insulation is displayed if either main gas or cavity wall
insulation was not selected in the previous question.

Solid wall insulation is displayed if either main gas or solid wall
insulation was not selected in the previous question.

It's this logic (from point d) that I have decided to use in the
template, because it actually checks whether the user has selected
cavity wall or solid wall insulation.